move /src/quickEvalCodeLensProvider.ts to /src/local-queries/quick-eval-code-lens-provider
